Moved common config to own folder
